backend/app/requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
gunicorn

social-auth-core
social-auth-app-django

drf_social_oauth2